refactor: split menu components
