+++ src/components/Header/Header.tsx
@@ -1,45 +1,13 @@
 'use client'
 
-import { faBars, faClose } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import cn from 'classnames'
 import Image from 'next/image'
 import Link from 'next/link'
-import { usePathname } from 'next/navigation'
-import { MouseEventHandler, useEffect, useRef, useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 
-import menuData, { MenuItem } from '@/components/Header/data'
+import menuData from '@/components/Header/data'
+import MenuLink from '@/components/Header/MenuLink'
+import MenuToggleButton from '@/components/Header/MenuToggleButton'
 import useClickOutside from '@/hooks/useClickOutside'
-
-interface MenuItemLink extends MenuItem {
-  onClick: MouseEventHandler<HTMLAnchorElement>
-}
-
-const isActiveLink = (menuPath: string, routePath: string | null) =>
-  menuPath.replace(/\/\d+/, '') === routePath?.replace(/\/\d+/, '')
-
-const MenuLink = ({ icon, href, onClick, text }: MenuItemLink) => {
-  const currentPath = usePathname()
-  const activeLink = isActiveLink(href, currentPath)
-
-  const classes =
-    'flex items-center px-5 py-3 rounded-md font-medium text-sm bg-gray-850 hover:text-white hover:bg-gradient-to-r from-brand-background-element to-rose-900'
-  const className = cn(classes, {
-    'bg-gradient-to-r text-white': activeLink,
-    'text-gray-400': !activeLink,
-  })
-
-  return (
-    <li className="px-2 xl:pl-5">
-      <Link className={className} href={href} onClick={onClick}>
-        <span className="mr-2 text-xl">
-          <FontAwesomeIcon className="w-4" icon={icon} />
-        </span>
-        {text}
-      </Link>
-    </li>
-  )
-}
 
 export default function Header() {
   const [menuOpen, setMenuOpen] = useState(false)
@@ -64,17 +32,11 @@ export default function Header() {
           <Image alt="logo.svg" src="/logo.svg" height={32} width={192} />
         </Link>
         <div className="lg:hidden">
-          <button
-            aria-label="Menu Button"
-            className="flex h-9 w-9 items-center justify-center rounded-full bg-brand-background-element"
+          <MenuToggleButton
             onClick={toggleMenuOpen}
+            open={menuOpen}
             ref={buttonRef}
-          >
-            <FontAwesomeIcon
-              className="text-2xl text-white"
-              icon={menuOpen ? faClose : faBars}
-            />
-          </button>
+          />
         </div>
       </div>
       <nav className={`${menuOpen ? 'block bg-black' : 'hidden lg:block'}`}>
